+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # EDGEGRID GOLANG RELEASE NOTES
 
+## 4.1.0 (Feb 27, 2023)
+
+### FEATURES/ENHANCEMENTS:
+
+* APPSEC
+  * Added following BotManager fields to GetExportConfigurationResponse
+    * BotManagement
+    * CustomBotCategories
+    * CustomDefinedBots
+    * CustomBotCategorySequence
+    * CustomClients
+    * ResponseActions
+    * AdvancedSettings
+  * Added AdvancedSettingsAttackPayloadLogging interface
+
+#### BUG FIXES:
+
+* Fix V4 of Edgegrid doesn't parse hostname ([#182](https://github.com/akamai/AkamaiOPEN-edgegrid-golang/pull/182))
+
 ## 4.0.0 (Jan 31, 2023)
 
 #### BREAKING CHANGES: